refactor(lib): promote thumbnail helper to lib/utilities
Move ThumbnailRef, the private thumbnailUrl, and resolveThumbnailUrl out
of lectures/lectureShape.ts into a lecture-agnostic
lib/utilities/thumbnailUrl.ts (project-structure rule 4 — cross-boundary
code belongs in src/lib/), so the upcoming meditation card shaper can
reuse it. lectureShape imports the helper; its resolveThumbnailUrl unit
cases move to tests/unit/thumbnail-url.spec.ts. No behavior change.

Refs #522
